Fix examples and doc-tests

+++ examples/opengl/opengl.rs
@@ -22,7 +22,11 @@ fn render_my_texture(
     let physical_size = hi_dpi_bounds.get_physical_size();
 
     let gl_context = info.window.get_gl_context();
-    let mut texture = Texture::new(gl_context.clone(), physical_size.width, physical_size.height);
+    let mut texture = Texture::new(
+        gl_context.clone(),
+        physical_size.width as usize,
+        physical_size.height as usize
+    );
 
     {
         let mut fb = FrameBuffer::new(&mut texture);
